requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langgraph>=0.1.0
chromadb>=0.4.0
PyPDF2>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.25.0
